adding save function to logreg class and renaming hyperopt.py
